cleanups suggested by @ikhoon
